+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to cssdb
 
+### Unreleased (patch)
+
+- Updated support data for `color-mix`
+- Updated support data for `gradients-interpolation-method`
+
 ### 7.5.1 (March 24, 2023)
 
 - Fixed potentially old generated files (patch)